+++ src/main/java/com/postech/mscreditcard/interfaces/ICustomerGateway.java
@@ -8,6 +8,6 @@ import java.util.List;
 public interface ICustomerGateway {
     CustomerDTO createCustomer(CustomerDTO customerDTO);
     CustomerDTO findByCpf(String cpf);
-    void deleteById(Long id);
+    void deleteById(Integer id);
     List<CustomerDTO> listAllCustomers();
 }
